backend/requirements.txt
fastapi[standard]
uvicorn
faiss-cpu
langchain
langchain_community
openai
tiktoken
python-dotenv